fixed some issues with paths and runcmd cernbox command

+++ python/smashbox/utilities/__init__.py
@@ -278,6 +278,9 @@ def oc_webdav_url(protocol='http',remote_folder="",user_num=None,webdav_endpoint
     else:
         password = config.oc_account_password
 
+    if platform.system() == "Windows":
+        remote_path = remote_path[:-1]
+
     return protocol + '://' + urllib.quote(username, safe='') + ':' + urllib.quote(password, safe='') + '@' + config.oc_server + '/' + remote_path
 
 
@@ -318,7 +321,10 @@ def run_ocsync(local_folder, remote_folder="", n=None, user_num=None):
 
     for i in range(n):
         t0 = datetime.datetime.now()
-        cmd = config.oc_sync_cmd+' '+local_folder+' '+oc_webdav_url('owncloud',remote_folder,user_num) + " >> "+config.rundir+"/%s-ocsync.step%02d.cnt%03d.log 2>&1"%(reflection.getProcessName(),current_step,ocsync_cnt[current_step])
+        if platform.system() == "Windows":
+           cmd = config.oc_sync_cmd  + " " + local_folder + " " +  oc_webdav_url('owncloud', remote_folder,user_num) + " >> " + config.rundir + "\%s-ocsync.step%02d.cnt%03d.log 2>&1" % (reflection.getProcessName(), current_step, ocsync_cnt[current_step])
+        else:
+           cmd = config.oc_sync_cmd,local_folder,oc_webdav_url('owncloud',remote_folder,user_num) + " >> "+config.rundir+"/%s-ocsync.step%02d.cnt%03d.log 2>&1"%(reflection.getProcessName(),current_step,ocsync_cnt[current_step])
         runcmd(cmd, ignore_exitcode=True)  # exitcode of ocsync is not reliable
         logger.info('sync cmd is: %s',cmd)
         logger.info('sync finished: %s',datetime.datetime.now()-t0)
@@ -355,13 +361,19 @@ def expect_webdav_exist(path, user_num=None):
 
 
 def webdav_delete(path, user_num=None):
-    runcmd('curl --verbose -k %s -X DELETE %s '%(config.get('curl_opts',''),oc_webdav_url(remote_folder=path, user_num=user_num)))
+    if platform.system() == "Windows":
+        webdav_delete_NEW(path, user_num)
+    else:
+       runcmd('curl --verbose -k %s -X DELETE %s '%(config.get('curl_opts',''),oc_webdav_url(remote_folder=path, user_num=user_num)))
     
 def webdav_mkcol(path, silent=False, user_num=None):
     out=""
-    if silent: # a workaround for super-verbose errors in case directory on the server already exists
-        out = "> /dev/null 2>&1"
-    runcmd('curl --verbose -k %s -X MKCOL %s %s'%(config.get('curl_opts',''),oc_webdav_url(remote_folder=path, user_num=user_num),out))
+    if platform.system() == "Windows":
+        webdav_mkcol_NEW(path,silent, user_num)
+    else:
+        if silent: # a workaround for super-verbose errors in case directory on the server already exists
+            out = "> /dev/null 2>&1"
+        runcmd('curl --verbose -k %s -X MKCOL %s %s'%(config.get('curl_opts',''),oc_webdav_url(remote_folder=path, user_num=user_num),out))
 
 
 # The two *_NEW functions below currently fail with:
@@ -390,8 +402,8 @@ def webdav_mkcol_NEW(path, silent=False, user_num=None):
 
 def runcmd(cmd,ignore_exitcode=False,echo=True,allow_stderr=True,shell=True,log_warning=True):
     logger.info('running %s', repr(cmd))
-
-    process = subprocess.Popen(cmd, shell=shell,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+    import sys
+    process = subprocess.Popen(cmd, shell=shell, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     stdout,stderr = process.communicate()
 
     if echo:
@@ -404,7 +416,7 @@ def runcmd(cmd,ignore_exitcode=False,echo=True,allow_stderr=True,shell=True,log_
                 logger.error("stderr: %s",stderr)
 
     if process.returncode != 0:
-        msg = "Non-zero exit code %d from command %s" % (ignore_exitcode,repr(cmd))
+        msg = "Non-zero exit code %s from command %s" % (ignore_exitcode,repr(cmd))
         if log_warning:
             logger.warning(msg)
         if not ignore_exitcode:
@@ -420,16 +432,16 @@ def sleep(n):
 
 ######## BASIC FILE AND DIRECTORY OPERATIONS
 
+import shutil
+
 def mkdir(d):
-    os.makedirs(d)
+    if not os.path.exists(d):
+       os.makedirs(d)
     return d
 
 
 def remove_tree(path):
-    if platform.system() == "Windows":
-       runcmd('rd / s / q ' + path)
-    else:
-       runcmd('rm -rf ' + path)
+    shutil.rmtree(path)
 
 def remove_file(path):
     logger.info('remove file %s',path)
@@ -444,21 +456,22 @@ def remove_file(path):
             raise
 
 def mv(a,b):
-    if platform.system() == "Windows":
-       runcmd('move %s %s' % (a, b))
-    else:
-       runcmd('mv %s %s' % (a, b))
+    shutil.move(a, b)
 
 def list_files(path,recursive=False):
-    if platform.system() == 'Darwin':
-        opts = ""
+    if platform.system() == 'Windows':
+        runcmd('dir /s /b ' + path)
     else:
-        opts = "--full-time"
+        if platform.system() == 'Darwin':
+            opts = ""
+        else:
+            opts = "--full-time"
 
-    if recursive:
-        runcmd('ls -lR %s %s'%(opts,path))
-    else:
-        runcmd('ls -lh %s %s'%(opts,path))
+        if recursive:
+            runcmd('ls -lR %s %s'%(opts,path))
+        else:
+            runcmd('ls -lh %s %s'%(opts,path))
+
 
 #http://stackoverflow.com/questions/377017/test-if-executable-exists-in-python#377028
 def which(program):
